Revert "hide homepage and redirect to blog for now"
This reverts commit 201312f2ba86fe2fed66b079d3ce72c53586ac11.
